chore(core): extract daemon client into a separate file
