+++ packages/workspace/src/core/project-graph/daemon/socket-utils.ts
@@ -1,9 +1,10 @@
 import { appRootPath } from '@nrwl/tao/src/utils/app-root';
 import { platform } from 'os';
 import { join, resolve } from 'path';
+import { unlinkSync } from 'fs';
 
 /**
- * For IPC with with the daemon server we use unix sockets or windows named pipes, depending on the user's operating system.
+ * For IPC with the daemon server we use unix sockets or windows named pipes, depending on the user's operating system.
  *
  * See https://nodejs.org/dist/latest-v14.x/docs/api/net.html#net_identifying_paths_for_ipc_connections for a full breakdown
  * of OS differences between Unix domain sockets and named pipes.
@@ -18,3 +19,9 @@ export const isWindows = platform() === 'win32';
 export const FULL_OS_SOCKET_PATH = isWindows
   ? '\\\\.\\pipe\\nx\\' + resolve(workspaceSocketPath)
   : resolve(workspaceSocketPath);
+
+export function killSocketOrPath(): void {
+  try {
+    unlinkSync(FULL_OS_SOCKET_PATH);
+  } catch {}
+}
